--- a/packages/contracts-bedrock/test/libraries/GasPayingToken.t.sol
+++ b/packages/contracts-bedrock/test/libraries/GasPayingToken.t.sol
@@ -7,9 +7,21 @@ import { Constants } from "src/libraries/Constants.sol";
 import { Test } from "forge-std/Test.sol";
 import { LibString } from "@solady/utils/LibString.sol";
 
+contract GasPayingToken_Harness {
+    function exposed_sanitize(string memory _str) public pure returns (bytes32) {
+        return GasPayingToken.sanitize(_str);
+    }
+}
+
 /// @title GasPayingToken_Roundtrip_Test
 /// @notice Tests the roundtrip of setting and getting the gas paying token.
 contract GasPayingToken_Roundtrip_Test is Test {
+    GasPayingToken_Harness harness;
+
+    function setUp() public {
+        harness = new GasPayingToken_Harness();
+    }
+
     /// @dev Test that the gas paying token correctly sets values in storage.
     function testFuzz_set_succeeds(address _token, uint8 _decimals, bytes32 _name, bytes32 _symbol) external {
         GasPayingToken.set(_token, _decimals, _name, _symbol);
@@ -104,7 +116,7 @@ contract GasPayingToken_Roundtrip_Test is Test {
 
         vm.expectRevert("GasPayingToken: string cannot be greater than 32 bytes");
 
-        GasPayingToken.sanitize(_str);
+        harness.exposed_sanitize(_str);
     }
 
     /// @dev Test that `sanitize` works as expected when the input string is empty.