Prepare the changes to move to foundry's version 1. @mds1 asked for a benchmark of building and testing, here you have the results on a M3 Max.

### After migration:

- just build-dev: 38.94s
- just build:  185.64s

(this includes the build as I am running just clean between benchmarks)

- just test-dev:  1043.79s (- compilation: 1004.85s)
- just test  917.71s (- compilation: 732.07s)

### Before migration:

- just build-dev:  39.15s
- just build: 188.48s 

(this includes the build as I am running just clean between benchmarks)

- just test-dev : 1011.86s (- compilation: 972.71s)
- just test: 935.23s (- compilation:  746.75s)